backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP and async
aiohttp>=3.8.0
aiofiles>=23.0.0

# RSS and web scraping
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration and utilities
pyyaml>=6.0
python-dateutil>=2.8.0

# Optional: Data processing
pandas>=2.0.0

